requirements.txt
flask
gunicorn
python-binance
binance-futures-connector

#pip3 install - r requirements.txt